presentation-api/controlling-ua/PresentationAvailability_multiple_display_onchange_manual.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Monitoring the list of available presentation displays for multiple displays.</title>
<link rel="author" title="mark a. foltz" href="https://github.com/mfoltzgoogle">
<link rel="help" href="http://w3c.github.io/presentation-api/#monitoring-the-list-of-available-presentation-displays">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="common.js"></script>

<p id="notice">Please wait for a moment...</p>
<p>The test passes if a "PASS" result appears.<br></p>

<script>
    // prevent the default timeout
    setup({explicit_timeout: true});

    const notice = document.getElementById('notice');

    promise_test(t => {
      // clean up the instruction notice when the test ends
      t.add_cleanup(() => {
        notice.parentNode.removeChild(notice);
      });

      // initialize a presentation request
      const request = new PresentationRequest(presentationUrls);

      let availability, timeout, numDisplays, expectedAvailability, finishTest;
      let allPromises = [];
      
      const removeAllDisplays = () => {
        allPromises.push(new Promise((resolve, reject) => {
          finishTest = resolve;
        }));
        
        numDisplays = 0;
        expectedAvailability = false;
        // show an instruction notice
        notice.textContent = 'Please make all presentation displays unavailable'
                          + ' and click this button: ';
        const button = document.createElement('button');
        button.textContent = 'Start Monitoring';
        button.onclick = getAvailability;
        notice.appendChild(button);
      };

      const addOneDisplay = () => {
        numDisplays++;
        expectedAvailability = true;
        // show an instruction notice
        notice.textContent = 'Please make a presentation display available'
                           + ' and click Start Monitoring, or click Finish if there are no'
                           + ' more displays:';

        let finishButton = document.getElementById('finishButton');
        if (!finishButton) {
          finishButton = document.createElement('button');
          finishButton.textContent = 'Finish';
          finishButton.onclick = finishTest;
          finishButton.id = 'finishButton';
         }
      };

      // check the event and its attributes
      const checkEvent = evt => {
        clearTimeout(timeout);
        timeout = undefined;

        assert_true(evt.isTrusted && !evt.bubbles && !evt.cancelable && evt instanceof Event, 'A simple event is fired.');
        assert_equals(evt.type, 'change', 'The event name is "change".');
        assert_equals(evt.target, availability, 'event.target is the presentation availability.');
      };

      const watchEvent = (obj, watcher, type) => {
        const watchHandler = new Promise(resolve => {
          obj['on' + type] = evt => { resolve(evt); };
        });
        return Promise.all([ watchHandler, watcher.wait_for(type) ]).then(results => {
          assert_equals(results[0], results[1], 'Both on' + type + ' and addEventListener pass the same event object.');
          return results[0];
        });
      };

      const getAvailability = () => {
        notice.textContent = 'Please wait for a moment... (It might take long time)';

        // set timeout to observe the presentation availability
        timeout = t.step_timeout(() => {
          t.force_timeout();
          t.done();
        }, 90000);

        allPromises.push(request.getAvailability().then(a => {
          availability = a;
          assert_equals(expectedAvailability,
                        availability.value, 'Value of the presentation availability is ' + expectedAvailability);
          if (numDisplays == 1) {
            // wait until a "change" event is fired
            let eventWatcher = new EventWatcher(t, availability, 'change');
            allPromises.push(watchEvent(availability, eventWatcher, 'change').then(checkEvent));
          }
        }));
      };

      removeAllDisplays();
      return Promise.all(allPromises);
    });
</script>
